feat: add user checks in watchlist controller

+++ app/Http/Controllers/WatchlistController.php
@@ -27,6 +27,9 @@ class WatchlistController extends Controller
      */
     public function store(WatchlistRequest $request)
     {
+        if ($request->user()->id != $request->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         $validated_data = (object) $request->validated();
         $watchlist = Watchlist::create(['name' => $validated_data->name, 'user_id' => $validated_data->user_id]);
         return response()->json(['watchlist' => $watchlist], 201);
@@ -38,8 +41,11 @@ class WatchlistController extends Controller
      * @param  \App\Models\Watchlist  $watchlist
      * @return \Illuminate\Http\Response
      */
-    public function show(Watchlist $watchlist)
+    public function show(Watchlist $watchlist, Request $request)
     {
+        if ($request->user()->id != $watchlist->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         return response()->json(["watchlist" => $watchlist->with('items')->get()], 200);
     }
 
@@ -52,6 +58,9 @@ class WatchlistController extends Controller
      */
     public function update(WatchlistRequest $request, Watchlist $watchlist)
     {
+        if ($request->user()->id != $watchlist->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         $watchlist->update($request->validated());
         return response()->json(['watchlist' => $watchlist], 200);
     }
@@ -62,15 +71,22 @@ class WatchlistController extends Controller
      * @param  \App\Models\Watchlist  $watchlist
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Watchlist $watchlist)
+    public function destroy(Request $request, Watchlist $watchlist)
     {
+        if ($request->user()->id != $watchlist->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         $watchlist->delete();
         return response()->json('', 204);
     }
 
     public function addItem(Request $request)
     {
+
         $watchlist = Watchlist::find($request->watchlist_id);
+        if ($request->user()->id != $watchlist->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         $item = Item::find($request->item_id);
         $item->watchlists()->sync([$item->id => [
             'rating' => 4.0,
@@ -81,6 +97,9 @@ class WatchlistController extends Controller
     public function removeItem(Request $request)
     {
         $watchlist = Watchlist::find($request->watchlist_id);
+        if ($request->user()->id != $watchlist->user_id) {
+            return response()->json(["error" => 'unauthorized'], 403);
+        }
         $item = Item::find($request->item_id);
         $item->watchlists()->detach($watchlist);
         return response()->json(['watchlist' => $watchlist], 200);
